Deserialization + Unit test.

+++ CS451_inventoryapp/src/com/example/cs451_inventorypackage/InventoryItem.java
@@ -46,7 +46,7 @@ public class InventoryItem implements Serializable {
 	}
 	
 	public boolean equals(InventoryItem i) {
-		return this.getId() == i.getId();
+		return this.getId().intValue() == i.getId().intValue();
 	}
 	public boolean equals(Integer itemId) {
 		return this.getId().equals(itemId);
